% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indek-cache-manager.R
\name{download_file}
\alias{download_file}
\title{The function 'download_file()', download the requested file from
the internet.}
\usage{
download_file(cache, url, filename)
}
\description{
## S3 methos for class 'indek.href'
}
