% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indek-page.R
\name{find_hrefs}
\alias{find_hrefs}
\title{'find_hrefs' search for resources whithin the links of a web page}
\usage{
find_hrefs(page, regex)
}
\arguments{
\item{regex}{a character vector with the regexs for matching.}

\item{x}{an object used to select a method}
}
\description{
## S3 method for class 'indek.page'
find_hrefs(x, regex)
}
