% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{indek.db.url}
\alias{indek.db.url}
\title{indec databases webpage url}
\format{An object of class \code{character} of length 1.}
\usage{
indek.db.url
}
\description{
indec databases webpage url
}
\keyword{datasets}
