% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indek-page-dbs.R
\name{indek.dbs}
\alias{indek.dbs}
\title{The function 'indek.dbs' is for fetching databases in the
indec page: https://www.indec.gob.ar/bases-de-datos.asp}
\usage{
indek.dbs(url = DBS_PAGE_URL, baseurl = DBS_BASE_URL,
  cachedir = DEFAULT_CACHE_DIR)
}
\description{
The function 'indek.dbs' is for fetching databases in the
indec page: https://www.indec.gob.ar/bases-de-datos.asp
}
\examples{
k <- indek.dbs()
eph <- find_hrefs(k, "EPH_usu_2_Trim_2018_txt.zip")[1]
x <- read_tables_zip(eph,1)[[1]]
summary(subset(x, PP08D1 != -9)$PP08D1)
}
