% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indek-page.R
\name{indek.page}
\alias{indek.page}
\title{The function 'indek.page()' creates pages, which are objects
that facilitate obtaining resources from web pages. It stores the
links that can be found in the page and permits the user to search
the required resources.}
\usage{
indek.page(url, baseurl, cachedir)
}
\arguments{
\item{url}{a character vector with the url of the web page}

\item{baseurl}{a character vector with the url of the web page.
For example, ohttps://www.indec.gob.ar.}

\item{cachedir}{a character vector with the local path to a
directory to be used as the cache.}
}
\description{
The function 'indek.page()' creates pages, which are objects
that facilitate obtaining resources from web pages. It stores the
links that can be found in the page and permits the user to search
the required resources.
}
\examples{
indek.page("https://www.indec.gob.ar/bases-de-datos.asp",
           "https://www.indec.gob.ar",
            "~/.indek.cache")
}
