% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indek-url.R
\name{read_tables_zip}
\alias{read_tables_zip}
\title{The function 'read_tables_zip' returns a list of data.frames
obtained from the zip file provided.}
\usage{
read_tables_zip(x, files)
}
\arguments{
\item{x}{the object that points to de zip file.}

\item{files}{the files within the zip file that are required}
}
\description{
It can be user to extract all the files or, if its content
is known, the required.
}
\details{
## S3 method for class 'indek.href'

## todo: S3 methos for url ("character vector")
}
